requirements.txt
geocoder==1.38.1
requests==2.31.0
Pillow==10.1.0
packaging==23.2
ttkbootstrap==1.10.1